Move NRQL test validation into the yaml definition
